% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmsSample.R
\name{emptySampleSet}
\alias{emptySampleSet}
\title{Generate an empty RTMS sample set}
\usage{
emptySampleSet(peaks)
}
\arguments{
\item{peaks}{A list of objects of class \code{rtmsPeak}}
}
\value{
An empty object of class \code{rtmsSampleSet}
}
\description{
Produces a sample set (of class \code{rtmsSampleSet}) with no samples but a
specific peaks attribute.  Useful for building a sample set one sample at
a time.
}
