% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleSpectrum}
\alias{exampleSpectrum}
\title{Simple Example Spectrum}
\format{
\subsection{\code{exampleSpectrum}}{

An object of class \code{rtmsSpectrum}
}
}
\source{
<St. Jude Children's Research Hospital>
}
\usage{
exampleSpectrum
}
\description{
A subset of an example spectrum depicting turnover from a protein substrate
(at 1530.8 m/z) to product (at 1516.8 m/z).  While the original spectrum
contained nearly 2 million measurements, this spectrum has been trimmed to
lie between 1300 m/z and 1600 m/z, with approximately 46000 measurements
included.
}
\keyword{datasets}
