% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmsBrukerBAFReader.R
\name{getBrukerBAFMetadata}
\alias{getBrukerBAFMetadata}
\title{Retrieve specific metadata values from a Bruker BAF file}
\usage{
getBrukerBAFMetadata(reader, names)
}
\arguments{
\item{reader}{An RTMS reader object of class \code{rtmsBrukerBAFReader}}

\item{names}{A character vector of metadata names}
}
\value{
A named list of values corresponding to the metadata values
specified.  All values will be returned as a string, including numeric
quantities (with units if appropriate).
}
\description{
Retrieves a list of specific metadata values (including instrument data,
acquisition parameters, processing and analysis directives, etc.) from a
Bruker single acquisition BAF directory (represented by an
\code{rtmsBrukerBAFReader} object).
}
