% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmsSample.R
\name{getSampleSet}
\alias{getSampleSet}
\title{Get an RTMS Sample Set}
\usage{
getSampleSet(x, peaks, ...)
}
\arguments{
\item{x}{The object from which the sample set should be retrieved}

\item{peaks}{A list of objects of class \code{rtmsPeak}}

\item{...}{Other possible arguments to specify a particular sample set to be
retrieved}
}
\value{
A sample set object of class \code{rtmsSampleSet}
}
\description{
Fetches an object of class \code{rtmsSampleSet} from the specified object.
}
