% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmsSpectrum.R
\name{getSpectrum}
\alias{getSpectrum}
\title{Get an RTMS Spectrum}
\usage{
getSpectrum(x, ...)
}
\arguments{
\item{x}{The object from which the spectrum should be retrieved}

\item{...}{Other possible arguments to specify a particular spectrum to be
retrieved}
}
\value{
An spectrum object of class \code{rtmsSpectrum}
}
\description{
Fetches an object of class \code{rtmsSpectrum} from the specified object.
}
