% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmsBrukerMCFReader.R
\name{reopen}
\alias{reopen}
\title{Reopen a reader object}
\usage{
reopen(x, ...)
}
\arguments{
\item{x}{The object to be re-opened}

\item{...}{Other possible arguments for specific object types}
}
\value{
An object of the same type as \code{x}
}
\description{
An S3 generic function for reopening reader objects that have been close
}
