% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmsSpectrum.R
\name{rtmsSpectrum}
\alias{rtmsSpectrum}
\title{Create a new RTMS spectrum}
\usage{
rtmsSpectrum(mz, intensity)
}
\arguments{
\item{mz}{A numeric vector of m/z values}

\item{intensity}{A numeric vector of intensity values}
}
\value{
An object of class \code{rtmsSpectrum} with the given m/z and intensity
values
}
\description{
Generates an RTMS spectrum object (of class \code{rtmsSpectrum}) from a given
vector of m/z and intensity values.
}
